requirements.txt
requests>=2.19.1
lxml>=4.2.5
pytest>=3.8.0
PyYAML>=3.12
flake8